ReactMapState を独立したカスタムフックに切り出した
